pyproject.toml
[tool.poetry]
name = "pgp-vlm-compression"
version = "0.0.1rc0"
description = "A project for data compression in multimodal models"
authors = ["Bardia Azizian <bardia_azizian@sfu.ca>"]
exclude = ["tests/*"]

[tool.poetry.dependencies]
python = "3.11.5"
matplotlib = "*"
mlflow-skinny = "*"
numpy = "<2.0.0"
pytorch-lightning = "^2.3.3"
sfu-torch-lib = "0.0.48"
tensorboardx = "*"
tensorboard = "*"
ipykernel = "*"
pandas = "*"
numba = "0.61.0"
defopt = "6.4.0"
ms-vlmeval = "0.0.18"
pillow-heif = "1.1.1"
bjontegaard = "1.3.0"
compressai = "1.2.8"
accelerate = "*"
bitsandbytes = "0.45.2"
decord = "*"
deepspeed = "0.16.3"
einops = "0.6.1"
einops-exts = "0.0.4"
huggingface_hub = "*"
imageio = "*"
opencv-python = "*"
orjson = "*"
peft = "0.10.0"
pycocoevalcap = "*"
pyyaml = "*"
scikit-learn = ">=1.2.2"
scipy = "*"
sentencepiece = "0.1.99"
shortuuid = "*"
termcolor = "*"
timm = "0.9.12"
tokenizers = "*"
torch = "2.6.0"
torchvision = ">=0.15"
tqdm = "*"
transformers = "4.48.0"
yacs = "*"
seaborn = "0.13.2"
datasets = "*"
open_clip_torch = "*"
setuptools = "*"

[tool.poetry.extras]
extras = []

[tool.poetry.group.dev.dependencies]
defopt = "*"
ruff = "*"
basedpyright = "*"
pytest = "*"

[tool.ruff]
line-length = 120
preview = true

[tool.ruff.format]
quote-style = "single"

[tool.pyright]
typeCheckingMode = "off"

[tool.isort]
line_length = 120
multi_line_output = 9

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
